wezterm: fix an issue where unzooming wouldn't resize the pane
The problem was that the pane resizes got optimized away because the
tab size doesn't change as part of the zoom operation.

We don't need to run the full resize logic for unzoom though, just
need to re-apply the size to the unzoomed pane.

We don't have a convenient way to do that, so we re-apply the sizes
to all of the panes in the tab.

refs: https://github.com/wez/wezterm/issues/157#issuecomment-701173561