windows: properly fallback to JetBrains Mono
The default font locator on Windows is font-loader.  That will always
return *something* that matched, even if it didn't match what we asked
for very well.  This is problematic if for example we asked for
"JetBrains Mono" and that isn't installed.  font-loader will return
the data for "Consolas".  That's great if there is no other fallback
mechanism, but it prevented us from searching our built-in fonts.

This commit parsed the returned font and uses the same font
family matching that we use for our own font directory searching
to see if it matched well enough.

This commit also tidies up that matching routine so that it more
carefully matches sub-family characteristics such as bold, italic
and bold-italic.