wezterm: don't give up if we run out of fallbacks
9d33073d7058f00d50b0ac12c647715296dbaa16 improved how we handle
the case where we run out of fallbacks, but with the unintended
consequence that we'd just give up on rendering the current frame,
which isn't super useful.

Since that commit landed we now bundle JetBrains Mono so that
means that we can guarantee to find our fallback glyph in case
we can't locate the font holding the requested glyph.

This commit causes us to unconditionally recurse and shape
the replacement character to use in this situation.

refs: #299