wezterm: add a linear scaling sampler for graphics
The fragment shader was set to use nearest texel sampling
because that made for sharper lines when rendering the
cursor bar or outlines.

This commit adds a supplemental sampler over the same
texture that is set to merge nearby texels.  We now
use this sampler for the background image and cells
marked as having full color textures (eg: emoji,
image protocol).

This makes the window background image look better
when it is scaled.  I'm not 100% sure about image
protocol cells, but I'm looking into that in general
at the moment.