wezterm: workaround appimage startup issue
AppImage passes fd 1023 opened on the mount point directory.
If we close this prior to exec, some magic in appimage decides
to unmount the fuse mount right before we're about to exec
the wezterm-gui binary from the appimage.

So, don't close stray fds when the cli spawns the gui when
we're an appimage.

See also: e831032c725ce2d4bc63f0bf1ecf33364cf72dac

closes: #419