ci: adjust continuous/nightly schedule
This commit reduces the cron schedule for nightly builds down from
hourly and to actually nightly (at ~3am).

The push-to-master flow now triggers effectively the same thing at push
time, so if anything, this will reduce the latency of the continuous
package build and deploy.

The push-to-master flow will now skip builds if it changes only the
docs.